refactor: change os structure
